Changed structure of app and corrected services
