+++ app/build.gradle
@@ -90,9 +90,9 @@ dependencies {
     implementation "org.koin:koin-androidx-ext:$koin_version"
 
     // Room database for cache and offline experience
-    //implementation "androidx.room:room-runtime:$room_version"
-    //implementation "androidx.room:room-ktx:$room_version"
-    //kapt "androidx.room:room-compiler:$room_version"
+    implementation "androidx.room:room-runtime:$room_version"
+    implementation "androidx.room:room-ktx:$room_version"
+    kapt "androidx.room:room-compiler:$room_version"
 
     // WorkManager for fetching data in background
     // implementation "androidx.work:work-runtime-ktx:$work_version"
@@ -104,5 +104,3 @@ dependencies {
 
 
 }
-    // Additional better logging for http request
-    // implementation 'com.squareup.okhttp3:logging-interceptor:4.6.0'
