+++ app/src/main/java/com/example/ing_app/network/comment/CommentService.kt
@@ -1,4 +1,4 @@
-package com.example.ing_app.network.Comment
+package com.example.ing_app.network.comment
 
 import com.example.ing_app.domain.Comment
 import kotlinx.coroutines.Deferred
@@ -8,4 +8,6 @@ import retrofit2.http.Query
 interface CommentService {
     @GET("/comments")
     fun getCommentsFromPost(@Query("postId") postId: Int): Deferred<List<Comment>>
+    @GET("/comments")
+    fun getComments(): Deferred<List<Comment>>
 }
