+++ app/src/main/java/com/example/ing_app/network/image/ImageApi.kt
@@ -1,4 +1,4 @@
-package com.example.ing_app.network.Post
+package com.example.ing_app.network.image
 
 import android.content.Context
 import com.example.ing_app.R
@@ -8,15 +8,10 @@ import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
-
-import timber.log.Timber
-import java.lang.Exception
 import java.util.concurrent.TimeUnit
 
-
-class PostApi(private val context: Context){
+class ImageApi (private val context: Context){
     private val BASE_URL = "https://jsonplaceholder.typicode.com"
-
 
     // Build the Moshi object that Retrofit will be using, making sure to add the Kotlin adapter for
     // full Kotlin compatibility.
@@ -36,9 +31,10 @@ class PostApi(private val context: Context){
         .addConverterFactory(MoshiConverterFactory.create(moshi))
         .addCallAdapterFactory(CoroutineCallAdapterFactory())
         .baseUrl(BASE_URL)
+        .client(okBuilder)
         .build()
 
-    fun getApiService(): PostService {
-            return retrofit.create(PostService::class.java)
+    fun getApiService(): ImageService {
+        return retrofit.create(ImageService::class.java)
     }
 }
