+++ app/src/main/java/com/example/ing_app/network/post/PostApi.kt
@@ -1,4 +1,4 @@
-package com.example.ing_app.network.Comment
+package com.example.ing_app.network.post
 
 import android.content.Context
 import com.example.ing_app.R
@@ -8,10 +8,13 @@ import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
+
 import java.util.concurrent.TimeUnit
 
-class CommentApi (private val context: Context){
+
+class PostApi(private val context: Context){
     private val BASE_URL = "https://jsonplaceholder.typicode.com"
+
 
     // Build the Moshi object that Retrofit will be using, making sure to add the Kotlin adapter for
     // full Kotlin compatibility.
@@ -31,10 +34,9 @@ class CommentApi (private val context: Context){
         .addConverterFactory(MoshiConverterFactory.create(moshi))
         .addCallAdapterFactory(CoroutineCallAdapterFactory())
         .baseUrl(BASE_URL)
-        .client(okBuilder)
         .build()
 
-    fun getApiService(): CommentService {
-        return retrofit.create(CommentService::class.java)
+    fun getApiService(): PostService {
+            return retrofit.create(PostService::class.java)
     }
 }
