+++ app/src/main/java/com/example/ing_app/network/post/PostService.kt
@@ -1,0 +1,10 @@
+package com.example.ing_app.network.post
+
+import com.example.ing_app.domain.Post
+import kotlinx.coroutines.Deferred
+import retrofit2.http.GET
+
+interface PostService {
+    @GET("/posts")
+    fun getPosts(): Deferred<List<Post>>
+}
