+++ app/src/main/java/com/example/ing_app/network/user/UserService.kt
@@ -1,4 +1,4 @@
-package com.example.ing_app.network.User
+package com.example.ing_app.network.user
 
 import com.example.ing_app.domain.User
 import kotlinx.coroutines.Deferred
@@ -8,4 +8,6 @@ import retrofit2.http.Path
 interface UserService {
     @GET("/users/{userId}")
     fun getUser(@Path("userId") userId : Int): Deferred<User>
+    @GET("/users")
+    fun getUsers(): Deferred<List<User>>
 }
